ll_web/requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.6
django-cron==0.6.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.5
